moving function to more logical location
