+++ src/common/util.rs
@@ -36,29 +36,3 @@ pub fn display_plan_with_partition_in_out(plan: &dyn ExecutionPlan) -> Result<St
     visit(plan, 0, &mut f)?;
     Ok(f)
 }
-
-/// Returns a boolean indicating if this stage can be divided into more than one task.
-///
-/// Some Plan nodes need to materialize all partitions inorder to execute such as
-/// NestedLoopJoinExec.   Rewriting the plan to accommodate dividing it into tasks
-/// would result in redundant work.
-///
-/// The plans we cannot split are:
-/// - NestedLoopJoinExec
-pub fn can_be_divided(plan: &Arc<dyn ExecutionPlan>) -> Result<bool> {
-    // recursively check to see if this stages plan contains a NestedLoopJoinExec
-    let mut has_unsplittable_plan = false;
-    let search = |f: &Arc<dyn ExecutionPlan>| {
-        if f.as_any()
-            .downcast_ref::<datafusion::physical_plan::joins::NestedLoopJoinExec>()
-            .is_some()
-        {
-            has_unsplittable_plan = true;
-            return Ok(TreeNodeRecursion::Stop);
-        }
-
-        Ok(TreeNodeRecursion::Continue)
-    };
-    plan.apply(search)?;
-    Ok(!has_unsplittable_plan)
-}
